requirements.txt
django==5.0.1
djangorestframework==3.14.0
redis==5.0.1
async-timeout==4.0.3